tm revamp: TestCheckMastership
Also fixed a race in shard sync.
