+++ go/vt/vttablet/tabletmanager/tm_init.go
@@ -448,6 +448,41 @@ func (tm *TabletManager) createKeyspaceShard(ctx context.Context) error {
 	return nil
 }
 
+func (tm *TabletManager) rebuildKeyspace(keyspace string, retryInterval time.Duration) {
+	var srvKeyspace *topodatapb.SrvKeyspace
+	defer func() {
+		log.Infof("Keyspace rebuilt: %v", keyspace)
+		tm.mutex.Lock()
+		defer tm.mutex.Unlock()
+		tm._srvKeyspace = srvKeyspace
+	}()
+
+	// RebuildKeyspace will fail until at least one tablet is up for every shard.
+	firstTime := true
+	var err error
+	for {
+		if !firstTime {
+			// If keyspace was rebuilt by someone else, we can just exit.
+			srvKeyspace, err = tm.TopoServer.GetSrvKeyspace(tm.BatchCtx, tm.tabletAlias.Cell, keyspace)
+			if err == nil {
+				return
+			}
+		}
+		err = topotools.RebuildKeyspace(tm.BatchCtx, logutil.NewConsoleLogger(), tm.TopoServer, keyspace, []string{tm.tabletAlias.Cell})
+		if err == nil {
+			srvKeyspace, err = tm.TopoServer.GetSrvKeyspace(tm.BatchCtx, tm.tabletAlias.Cell, keyspace)
+			if err == nil {
+				return
+			}
+		}
+		if firstTime {
+			log.Warningf("rebuildKeyspace failed, will retry every %v: %v", retryInterval, err)
+		}
+		firstTime = false
+		time.Sleep(retryInterval)
+	}
+}
+
 func (tm *TabletManager) checkMastership(ctx context.Context) error {
 	tm.mutex.Lock()
 	si := tm._shardInfo
@@ -529,6 +564,22 @@ func (tm *TabletManager) checkMysql(ctx context.Context) error {
 		}
 	}
 	return nil
+}
+
+func (tm *TabletManager) findMysqlPort(retryInterval time.Duration) {
+	for {
+		time.Sleep(retryInterval)
+		mport, err := tm.MysqlDaemon.GetMysqlPort()
+		if err != nil {
+			continue
+		}
+		log.Infof("Identified mysql port: %v", mport)
+		tm.pubMu.Lock()
+		tm.tablet.MysqlPort = mport
+		tm.pubMu.Unlock()
+		tm.publishState(tm.BatchCtx)
+		return
+	}
 }
 
 func (tm *TabletManager) initTablet(ctx context.Context) error {
