+++ go/vt/vttablet/tabletmanager/tm_init_test.go
@@ -17,15 +17,12 @@ limitations under the License.
 package tabletmanager
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/net/context"
-	"vitess.io/vitess/go/history"
-	"vitess.io/vitess/go/mysql/fakesqldb"
 	"vitess.io/vitess/go/sync2"
 	"vitess.io/vitess/go/vt/dbconfigs"
 	"vitess.io/vitess/go/vt/logutil"
@@ -35,7 +32,6 @@ import (
 	"vitess.io/vitess/go/vt/topo"
 	"vitess.io/vitess/go/vt/topo/memorytopo"
 	"vitess.io/vitess/go/vt/topotools"
-	"vitess.io/vitess/go/vt/vttablet/tabletmanager/vreplication"
 	"vitess.io/vitess/go/vt/vttablet/tabletservermock"
 )
 
@@ -124,7 +120,8 @@ func TestStartCreateKeyspaceShard(t *testing.T) {
 	ctx := context.Background()
 	cell := "cell1"
 	ts := memorytopo.NewServer(cell)
-	_ = newTestTM(t, ts, 1, "ks", "0")
+	tm := newTestTM(t, ts, 1, "ks", "0")
+	defer tm.Stop()
 
 	_, err := ts.GetShard(ctx, "ks", "0")
 	require.NoError(t, err)
@@ -139,7 +136,8 @@ func TestStartCreateKeyspaceShard(t *testing.T) {
 	// keyspace-shard already created.
 	_, err = ts.GetOrCreateShard(ctx, "ks1", "0")
 	require.NoError(t, err)
-	_ = newTestTM(t, ts, 2, "ks1", "0")
+	tm = newTestTM(t, ts, 2, "ks1", "0")
+	defer tm.Stop()
 	_, err = ts.GetShard(ctx, "ks1", "0")
 	require.NoError(t, err)
 	ensureSrvKeyspace(t, ts, cell, "ks1")
@@ -152,7 +150,8 @@ func TestStartCreateKeyspaceShard(t *testing.T) {
 	require.NoError(t, err)
 	err = topotools.RebuildKeyspace(ctx, logutil.NewConsoleLogger(), ts, "ks2", []string{cell})
 	require.NoError(t, err)
-	_ = newTestTM(t, ts, 3, "ks2", "0")
+	tm = newTestTM(t, ts, 3, "ks2", "0")
+	defer tm.Stop()
 	_, err = ts.GetShard(ctx, "ks2", "0")
 	require.NoError(t, err)
 	_, err = ts.GetSrvKeyspace(context.Background(), cell, "ks2")
@@ -168,7 +167,8 @@ func TestStartCreateKeyspaceShard(t *testing.T) {
 	require.NoError(t, err)
 	err = ts.RebuildSrvVSchema(ctx, []string{cell})
 	require.NoError(t, err)
-	_ = newTestTM(t, ts, 4, "ks3", "0")
+	tm = newTestTM(t, ts, 4, "ks3", "0")
+	defer tm.Stop()
 	_, err = ts.GetShard(ctx, "ks3", "0")
 	require.NoError(t, err)
 	_, err = ts.GetSrvKeyspace(context.Background(), cell, "ks3")
@@ -193,12 +193,139 @@ func ensureSrvKeyspace(t *testing.T, ts *topo.Server, cell, keyspace string) {
 	assert.True(t, found)
 }
 
+func TestCheckMastership(t *testing.T) {
+	defer func(saved time.Duration) { rebuildKeyspaceRetryInterval = saved }(rebuildKeyspaceRetryInterval)
+	rebuildKeyspaceRetryInterval = 10 * time.Millisecond
+
+	ctx := context.Background()
+	cell := "cell1"
+	ts := memorytopo.NewServer(cell)
+	alias := &topodatapb.TabletAlias{
+		Cell: "cell1",
+		Uid:  1,
+	}
+
+	// 1. Initialize the tablet as REPLICA.
+	// This will create the respective topology records.
+	tm := newTestTM(t, ts, 1, "ks", "0")
+	tablet := tm.Tablet()
+	ensureSrvKeyspace(t, ts, cell, "ks")
+	ti, err := ts.GetTablet(ctx, alias)
+	require.NoError(t, err)
+	assert.Equal(t, topodatapb.TabletType_REPLICA, ti.Type)
+	tm.Stop()
+
+	// 2. Update shard's master to our alias, then try to init again.
+	// (This simulates the case where the MasterAlias in the shard record says
+	// that we are the master but the tablet record says otherwise. In that case,
+	// we assume we are not the MASTER.)
+	_, err = ts.UpdateShardFields(ctx, "ks", "0", func(si *topo.ShardInfo) error {
+		si.MasterAlias = alias
+		return nil
+	})
+	require.NoError(t, err)
+	err = tm.Start(tablet)
+	require.NoError(t, err)
+	ti, err = ts.GetTablet(ctx, alias)
+	require.NoError(t, err)
+	assert.Equal(t, topodatapb.TabletType_REPLICA, ti.Type)
+	ter0 := ti.GetMasterTermStartTime()
+	assert.True(t, ter0.IsZero())
+	tm.Stop()
+
+	// 3. Delete the tablet record. The shard record still says that we are the
+	// MASTER. Since it is the only source, we assume that its information is
+	// correct and start as MASTER.
+	err = ts.DeleteTablet(ctx, alias)
+	require.NoError(t, err)
+	err = tm.Start(tablet)
+	require.NoError(t, err)
+	ti, err = ts.GetTablet(ctx, alias)
+	require.NoError(t, err)
+	assert.Equal(t, topodatapb.TabletType_MASTER, ti.Type)
+	ter1 := ti.GetMasterTermStartTime()
+	tm.Stop()
+
+	// 4. Fix the tablet record to agree that we're master.
+	// Shard and tablet record are in sync now and we assume that we are actually
+	// the MASTER.
+	ti.Type = topodatapb.TabletType_MASTER
+	err = ts.UpdateTablet(ctx, ti)
+	require.NoError(t, err)
+	err = tm.Start(tablet)
+	require.NoError(t, err)
+	ti, err = ts.GetTablet(ctx, alias)
+	require.NoError(t, err)
+	assert.Equal(t, topodatapb.TabletType_MASTER, ti.Type)
+	ter2 := ti.GetMasterTermStartTime()
+	assert.Equal(t, ter1, ter2)
+	tm.Stop()
+
+	// 5. Subsequent inits will still start the vttablet as MASTER.
+	err = tm.Start(tablet)
+	require.NoError(t, err)
+	ti, err = ts.GetTablet(ctx, alias)
+	require.NoError(t, err)
+	assert.Equal(t, topodatapb.TabletType_MASTER, ti.Type)
+	ter3 := ti.GetMasterTermStartTime()
+	assert.Equal(t, ter1, ter3)
+	tm.Stop()
+
+	// 6. If the shard record shows a different master with an older
+	// timestamp, we take over mastership.
+	otherAlias := &topodatapb.TabletAlias{
+		Cell: "cell1",
+		Uid:  2,
+	}
+	_, err = ts.UpdateShardFields(ctx, "ks", "0", func(si *topo.ShardInfo) error {
+		si.MasterAlias = otherAlias
+		si.MasterTermStartTime = logutil.TimeToProto(ter1.Add(-10 * time.Second))
+		return nil
+	})
+	require.NoError(t, err)
+	err = tm.Start(tablet)
+	require.NoError(t, err)
+	ti, err = ts.GetTablet(ctx, alias)
+	require.NoError(t, err)
+	assert.Equal(t, topodatapb.TabletType_MASTER, ti.Type)
+	ter4 := ti.GetMasterTermStartTime()
+	assert.Equal(t, ter1, ter4)
+	tm.Stop()
+}
+
+func TestFindMysqlPort(t *testing.T) {
+	defer func(saved time.Duration) { mysqlPortRetryInterval = saved }(mysqlPortRetryInterval)
+	mysqlPortRetryInterval = 1 * time.Millisecond
+
+	ctx := context.Background()
+	tm := createTestTM(ctx, t, nil)
+	err := tm.checkMysql(ctx)
+	require.NoError(t, err)
+	ttablet, err := tm.TopoServer.GetTablet(ctx, tm.tabletAlias)
+	require.NoError(t, err)
+	assert.Equal(t, ttablet.MysqlPort, int32(0))
+
+	tm.pubMu.Lock()
+	tm.MysqlDaemon.(*fakemysqldaemon.FakeMysqlDaemon).MysqlPort.Set(3306)
+	tm.pubMu.Unlock()
+	for i := 0; i < 10; i++ {
+		ttablet, err := tm.TopoServer.GetTablet(ctx, tm.tabletAlias)
+		require.NoError(t, err)
+		if ttablet.MysqlPort == 3306 {
+			return
+		}
+		time.Sleep(5 * time.Millisecond)
+	}
+	assert.Fail(t, "mysql port was not updated")
+}
+
 // Init tablet fixes replication data when safe
 func TestStartFixesReplicationData(t *testing.T) {
 	ctx := context.Background()
 	cell := "cell1"
 	ts := memorytopo.NewServer(cell, "cell2")
 	tm := newTestTM(t, ts, 1, "ks", "0")
+	defer tm.Stop()
 	tabletAlias := tm.tabletAlias
 
 	sri, err := ts.GetShardReplication(ctx, cell, "ks", "0")
@@ -230,6 +357,7 @@ func TestStartDoesNotUpdateReplicationDataForTabletInWrongShard(t *testing.T) {
 	ctx := context.Background()
 	ts := memorytopo.NewServer("cell1", "cell2")
 	tm := newTestTM(t, ts, 1, "ks", "0")
+	defer tm.Stop()
 
 	tabletAliases, err := ts.FindAllTabletAliasesInShard(ctx, "ks", "0")
 	require.NoError(t, err)
@@ -245,225 +373,13 @@ func TestStartDoesNotUpdateReplicationDataForTabletInWrongShard(t *testing.T) {
 	assert.Equal(t, 0, len(tablets))
 }
 
-// TestStart will test the Start code creates / updates the
-// tablet node correctly. Note we modify global parameters (the flags)
-// so this has to be in one test.
-func TestStart(t *testing.T) {
-	ctx := context.Background()
-	ts := memorytopo.NewServer("cell1", "cell2")
-	tabletAlias := &topodatapb.TabletAlias{
-		Cell: "cell1",
-		Uid:  1,
-	}
-	db := fakesqldb.New(t)
-	defer db.Close()
-
-	// start with a tablet record that doesn't exist
-	port := int32(1234)
-	gRPCPort := int32(3456)
-	mysqlDaemon := fakemysqldaemon.NewFakeMysqlDaemon(db)
-	tm := &TabletManager{
-		TopoServer:     ts,
-		tabletAlias:    tabletAlias,
-		MysqlDaemon:    mysqlDaemon,
-		DBConfigs:      &dbconfigs.DBConfigs{},
-		VREngine:       vreplication.NewTestEngine(nil, "", nil, nil, "", nil),
-		BatchCtx:       ctx,
-		History:        history.New(historyLength),
-		baseTabletType: topodatapb.TabletType_REPLICA,
-		_healthy:       fmt.Errorf("healthcheck not run yet"),
-	}
-
-	// 1. Initialize the tablet as REPLICA.
-	// This will create the respective topology records.
-	// We use a capitalized shard name here, to make sure the
-	// Keyrange computation works, fills in the KeyRange, and converts
-	// it to lower case.
-	*tabletHostname = "localhost"
-	*initKeyspace = "test_keyspace"
-	*initShard = "0"
-	*initTabletType = "replica"
-	tabletAlias = &topodatapb.TabletAlias{
-		Cell: "cell1",
-		Uid:  2,
-	}
-
-	_, err := tm.TopoServer.GetSrvKeyspace(ctx, "cell1", "test_keyspace")
-	switch {
-	case topo.IsErrType(err, topo.NoNode):
-		// srvKeyspace should not be when tablets haven't been registered to this cell
-	default:
-		t.Fatalf("GetSrvKeyspace failed: %v", err)
-	}
-
-	tm.tabletAlias = tabletAlias
-
-	tablet, err := BuildTabletFromInput(tabletAlias, port, gRPCPort)
-	require.NoError(t, err)
-	tm.tablet = tablet
-	err = tm.createKeyspaceShard(context.Background())
-	require.NoError(t, err)
-	err = tm.initTablet(context.Background())
-	require.NoError(t, err)
-
-	ti, err := ts.GetTablet(ctx, tabletAlias)
-	if err != nil {
-		t.Fatalf("GetTablet failed: %v", err)
-	}
-	if ti.Type != topodatapb.TabletType_REPLICA {
-		t.Errorf("wrong tablet type: %v", ti.Type)
-	}
-	if ti.Hostname != "localhost" {
-		t.Errorf("wrong hostname for tablet: %v", ti.Hostname)
-	}
-	if ti.PortMap["vt"] != port {
-		t.Errorf("wrong port for tablet: %v", ti.PortMap["vt"])
-	}
-	if ti.PortMap["grpc"] != gRPCPort {
-		t.Errorf("wrong gRPC port for tablet: %v", ti.PortMap["grpc"])
-	}
-	if ti.Shard != "0" {
-		t.Errorf("wrong shard for tablet: %v", ti.Shard)
-	}
-	if ti.KeyRange != nil {
-		t.Errorf("wrong KeyRange for tablet: %v", ti.KeyRange)
-	}
-	if got := tm.masterTermStartTime(); !got.IsZero() {
-		t.Fatalf("REPLICA tablet should not have a MasterTermStartTime set: %v", got)
-	}
-
-	// 2. Update shard's master to our alias, then try to init again.
-	// (This simulates the case where the MasterAlias in the shard record says
-	// that we are the master but the tablet record says otherwise. In that case,
-	// we assume we are not the MASTER.)
-	_, err = tm.TopoServer.UpdateShardFields(ctx, "test_keyspace", "0", func(si *topo.ShardInfo) error {
-		si.MasterAlias = tabletAlias
-		return nil
-	})
-	if err != nil {
-		t.Fatalf("UpdateShardFields failed: %v", err)
-	}
-
-	tablet, err = BuildTabletFromInput(tabletAlias, port, gRPCPort)
-	require.NoError(t, err)
-	tm.tablet = tablet
-	err = tm.createKeyspaceShard(context.Background())
-	require.NoError(t, err)
-	err = tm.initTablet(context.Background())
-	require.NoError(t, err)
-
-	ti, err = ts.GetTablet(ctx, tabletAlias)
-	if err != nil {
-		t.Fatalf("GetTablet failed: %v", err)
-	}
-	// It should still be replica, because the tablet record doesn't agree.
-	if ti.Type != topodatapb.TabletType_REPLICA {
-		t.Errorf("wrong tablet type: %v", ti.Type)
-	}
-	if got := tm.masterTermStartTime(); !got.IsZero() {
-		t.Fatalf("REPLICA tablet should not have a masterTermStartTime set: %v", got)
-	}
-
-	// 3. Delete the tablet record. The shard record still says that we are the
-	// MASTER. Since it is the only source, we assume that its information is
-	// correct and start as MASTER.
-	if err := ts.DeleteTablet(ctx, tabletAlias); err != nil {
-		t.Fatalf("DeleteTablet failed: %v", err)
-	}
-
-	tablet, err = BuildTabletFromInput(tabletAlias, port, gRPCPort)
-	require.NoError(t, err)
-	tm.tablet = tablet
-	err = tm.createKeyspaceShard(context.Background())
-	require.NoError(t, err)
-	err = tm.checkMastership(ctx)
-	require.NoError(t, err)
-	err = tm.initTablet(context.Background())
-	require.NoError(t, err)
-
-	ti, err = ts.GetTablet(ctx, tabletAlias)
-	if err != nil {
-		t.Fatalf("GetTablet failed: %v", err)
-	}
-	if ti.Type != topodatapb.TabletType_MASTER {
-		t.Errorf("wrong tablet type: %v", ti.Type)
-	}
-	ter1 := ti.GetMasterTermStartTime()
-	if ter1.IsZero() {
-		t.Fatalf("MASTER tablet should have a masterTermStartTime set")
-	}
-
-	// 4. Fix the tablet record to agree that we're master.
-	// Shard and tablet record are in sync now and we assume that we are actually
-	// the MASTER.
-	ti.Type = topodatapb.TabletType_MASTER
-	if err := ts.UpdateTablet(ctx, ti); err != nil {
-		t.Fatalf("UpdateTablet failed: %v", err)
-	}
-
-	tablet, err = BuildTabletFromInput(tabletAlias, port, gRPCPort)
-	require.NoError(t, err)
-	tm.tablet = tablet
-	err = tm.createKeyspaceShard(context.Background())
-	require.NoError(t, err)
-	err = tm.checkMastership(ctx)
-	require.NoError(t, err)
-	err = tm.initTablet(context.Background())
-	require.NoError(t, err)
-
-	ti, err = ts.GetTablet(ctx, tabletAlias)
-	if err != nil {
-		t.Fatalf("GetTablet failed: %v", err)
-	}
-	if ti.Type != topodatapb.TabletType_MASTER {
-		t.Errorf("wrong tablet type: %v", ti.Type)
-	}
-	ter2 := ti.GetMasterTermStartTime()
-	if ter2.IsZero() || !ter2.Equal(ter1) {
-		t.Fatalf("After a restart, masterTermStartTime must be equal to the previous time saved in the tablet record. Previous timestamp: %v current timestamp: %v", ter1, ter2)
-	}
-
-	// 5. Subsequent inits will still start the vttablet as MASTER.
-	// (Also check db name override and tags here.)
-	*initDbNameOverride = "DBNAME"
-	initTags.Set("aaa:bbb")
-
-	tablet, err = BuildTabletFromInput(tabletAlias, port, gRPCPort)
-	require.NoError(t, err)
-	tm.tablet = tablet
-	err = tm.createKeyspaceShard(context.Background())
-	require.NoError(t, err)
-	err = tm.checkMastership(ctx)
-	require.NoError(t, err)
-	err = tm.initTablet(context.Background())
-	require.NoError(t, err)
-
-	ti, err = ts.GetTablet(ctx, tabletAlias)
-	if err != nil {
-		t.Fatalf("GetTablet failed: %v", err)
-	}
-	if ti.Type != topodatapb.TabletType_MASTER {
-		t.Errorf("wrong tablet type: %v", ti.Type)
-	}
-	if ti.DbNameOverride != "DBNAME" {
-		t.Errorf("wrong tablet DbNameOverride: %v", ti.DbNameOverride)
-	}
-	if len(ti.Tags) != 1 || ti.Tags["aaa"] != "bbb" {
-		t.Errorf("wrong tablet tags: %v", ti.Tags)
-	}
-	ter3 := ti.GetMasterTermStartTime()
-	if ter3.IsZero() || !ter3.Equal(ter2) {
-		t.Fatalf("After a restart, masterTermStartTime must be set to the previous time saved in the tablet record. Previous timestamp: %v current timestamp: %v", ter2, ter3)
-	}
-}
-
 func newTestTM(t *testing.T, ts *topo.Server, uid int, keyspace, shard string) *TabletManager {
 	t.Helper()
 	tablet := newTestTablet(t, uid, keyspace, shard)
 	tm := &TabletManager{
 		BatchCtx:            context.Background(),
 		TopoServer:          ts,
-		MysqlDaemon:         &fakemysqldaemon.FakeMysqlDaemon{MysqlPort: sync2.NewAtomicInt32(-1)},
+		MysqlDaemon:         &fakemysqldaemon.FakeMysqlDaemon{MysqlPort: sync2.NewAtomicInt32(1)},
 		DBConfigs:           &dbconfigs.DBConfigs{},
 		QueryServiceControl: tabletservermock.NewController(),
 	}
